chore: move AOSCC event list to events page
